+++ wp-php-console.php
@@ -9,7 +9,7 @@
  * Plugin Name:       WP PHP Console
  * Plugin URI:        https://github.com/nekojira/wp-php-console/
  * Description:       An implementation of PHP Console for WordPress. Easily debug and trace PHP errors and warnings from your Chrome dev tools console using a Google Chrome extension.
- * Version:           1.3.3
+ * Version:           1.3.5
  * Author:            nekojira
  * Author URI:        https://github.com/nekojira/
  * License:           GPL-2.0+
@@ -18,7 +18,28 @@
  * Domain Path:       /languages
  */
 
-if ( ! defined( 'WPINC' ) ) die;
+if ( ! defined( 'WPINC' ) ) {
+	exit; // Exit if accessed directly
+}
+
+/**
+ * WP PHP Console requires PHP 5.4.0 minimum.
+ * WordPress supports 5.2.4 and recommends 5.4.0.
+ * @link https://make.wordpress.org/plugins/2015/06/05/policy-on-php-versions/
+ */
+if ( version_compare( PHP_VERSION, '5.4.0', '<') ) {
+	add_action( 'admin_notices',
+		function() {
+			echo '<div class="error"><p>'.
+			     sprintf( __( "WP PHP Console requires PHP 5.4 or above to function properly. Detected PHP version on your server is %s. Please upgrade PHP to activate WP PHP Console or remove the plugin.", 'wp-php-console' ), phpversion() ? phpversion() : '`undefined`' ) .
+			     '</p></div>';
+			if ( isset( $_GET['activate'] ) ) {
+				unset( $_GET['activate'] );
+			}
+		}
+	);
+	return;
+}
 
 /**
  * Include PhpConsole server library.